chore(insights): extract hog-charts Chart.tsx hooks and crosshair util
Move margin layout, y-tick formatter resolution, stable-resolveValue, and the
crosshair-composed drawHover out of core/Chart.tsx so the component reads as
"wire hooks together → render canvases + overlay". Behaviour preserved:
the drawHover ref + getter indirection keeps composedDrawHover stable across
parent drawHover identity changes.
