+++ frontend/src/lib/hog-charts/core/Chart.tsx
@@ -1,21 +1,20 @@
-import * as d3 from 'd3'
 import React, { useCallback, useMemo, useRef } from 'react'
 
-import { AxisLabels, measureLabelWidth } from '../overlays/AxisLabels'
+import { AxisLabels } from '../overlays/AxisLabels'
 import { DefaultTooltip } from '../overlays/DefaultTooltip'
 import { Tooltip } from '../overlays/Tooltip'
-import { drawCrosshair } from './canvas-renderer'
+import { composeDrawHoverWithCrosshair } from './canvas-renderer'
 import { ChartHoverContext, ChartLayoutContext } from './chart-context'
 import type { ChartHoverContextValue, ChartLayoutContextValue } from './chart-context'
 import { useChartCanvas } from './hooks/useChartCanvas'
 import { useChartDraw } from './hooks/useChartDraw'
 import { useChartInteraction } from './hooks/useChartInteraction'
-import { autoFormatYTick, seriesValueRange } from './scales'
-import { DEFAULT_Y_AXIS_ID } from './types'
+import { useChartMargins } from './hooks/useChartMargins'
+import { useResolvedYFormatters } from './hooks/useResolvedYFormatters'
+import { useStableResolveValue } from './hooks/useStableResolveValue'
 import type {
     ChartConfig,
     ChartDrawArgs,
-    ChartMargins,
     ChartScales,
     ChartTheme,
     CreateScalesFn,
@@ -25,6 +24,8 @@ import type {
     Series,
     TooltipContext,
 } from './types'
+
+export { DEFAULT_MARGINS } from './hooks/useChartMargins'
 
 const OVERLAY_STYLE: React.CSSProperties = {
     position: 'absolute',
@@ -80,8 +81,6 @@ export interface ChartProps<Meta = unknown> {
     resolveValue?: ResolveValueFn
 }
 
-export const DEFAULT_MARGINS: ChartMargins = { top: 16, right: 16, bottom: 32, left: 48 }
-
 export function Chart<Meta = unknown>({
     series,
     labels,
@@ -110,68 +109,7 @@ export function Chart<Meta = unknown>({
         placement: tooltipPlacement = 'follow-data',
     } = tooltipConfig ?? {}
 
-    const hasMultipleAxes = useMemo(() => {
-        const axisIds = new Set(
-            series.filter((s) => !s.visibility?.excluded).map((s) => s.yAxisId ?? DEFAULT_Y_AXIS_ID)
-        )
-        return axisIds.size > 1
-    }, [series])
-
-    const yLabelWidth = useMemo<number>(() => {
-        if (hideYAxis) {
-            return 0
-        }
-        const range = seriesValueRange(series)
-        if (range.count === 0) {
-            return 0
-        }
-        const min = range.min > 0 ? 0 : range.min
-        const max = range.max < 0 ? 0 : range.max
-        const ticks = d3.scaleLinear().domain([min, max]).nice(6).ticks(6)
-        if (ticks.length === 0) {
-            return 0
-        }
-        const domainMax = Math.max(...ticks.map((t) => Math.abs(t)))
-        const formatter = yTickFormatter ?? ((v: number) => autoFormatYTick(v, domainMax))
-        let widest = 0
-        for (const t of ticks) {
-            widest = Math.max(widest, measureLabelWidth(formatter(t)))
-        }
-        return widest
-    }, [series, yTickFormatter, hideYAxis])
-
-    const xLabelHalfWidth = useMemo<number>(() => {
-        if (hideXAxis || labels.length === 0) {
-            return 0
-        }
-        let widest = 0
-        for (let i = 0; i < labels.length; i++) {
-            const text = xTickFormatter ? xTickFormatter(labels[i], i) : labels[i]
-            if (text === null) {
-                continue
-            }
-            widest = Math.max(widest, measureLabelWidth(text))
-        }
-        return Math.ceil(widest / 2)
-    }, [labels, xTickFormatter, hideXAxis])
-
-    const margins = useMemo<ChartMargins>(() => {
-        const m = { ...DEFAULT_MARGINS }
-        if (hideXAxis) {
-            m.bottom = 8
-        }
-        if (hideYAxis) {
-            m.left = 8
-        } else {
-            m.left = Math.max(20, Math.ceil(yLabelWidth) + 12, xLabelHalfWidth + 4)
-        }
-        if (hasMultipleAxes && !hideYAxis) {
-            m.right = Math.max(48, xLabelHalfWidth + 4)
-        } else {
-            m.right = Math.max(DEFAULT_MARGINS.right, xLabelHalfWidth + 4)
-        }
-        return m
-    }, [hideXAxis, hideYAxis, hasMultipleAxes, yLabelWidth, xLabelHalfWidth])
+    const margins = useChartMargins({ series, labels, hideXAxis, hideYAxis, xTickFormatter, yTickFormatter })
 
     const { canvasRef, overlayCanvasRef, wrapperRef, dimensions, ctx, overlayCtx } = useChartCanvas({ margins })
 
@@ -191,27 +129,7 @@ export function Chart<Meta = unknown>({
         return createScalesFn(coloredSeries, labels, dimensions)
     }, [coloredSeries, labels, dimensions, createScalesFn])
 
-    const resolvedYFormatter = useMemo(() => {
-        if (yTickFormatter) {
-            return yTickFormatter
-        }
-        const ticks = scales?.yTicks() ?? []
-        const domainMax = ticks.length > 0 ? Math.abs(Math.max(...ticks)) : 1
-        return (v: number) => autoFormatYTick(v, domainMax)
-    }, [yTickFormatter, scales])
-
-    const resolvedYRightFormatter = useMemo(() => {
-        if (yTickFormatter) {
-            return yTickFormatter
-        }
-        const rightAxis = scales?.yAxes && Object.values(scales.yAxes).find((a) => a.position === 'right')
-        if (!rightAxis) {
-            return undefined
-        }
-        const ticks = rightAxis.ticks()
-        const domainMax = ticks.length > 0 ? Math.abs(Math.max(...ticks)) : 1
-        return (v: number) => autoFormatYTick(v, domainMax)
-    }, [yTickFormatter, scales])
+    const { left: resolvedYFormatter, right: resolvedYRightFormatter } = useResolvedYFormatters(scales, yTickFormatter)
 
     const { hoverIndex, tooltipCtx, handlers } = useChartInteraction<Meta>({
         scales,
@@ -226,27 +144,13 @@ export function Chart<Meta = unknown>({
         resolveValue,
     })
 
-    // Compose the chart-type's drawHover with a crosshair pass so per-mousemove
-    // hover indication stays entirely on the canvas — DOM-based overlays would
-    // force per-event style invalidation/layout that scales badly with chart
-    // content size. Crosshair drawn first so highlight rings render on top.
-    //
-    // drawHover is held via a ref so composedDrawHover stays referentially stable
-    // even when the parent recreates drawHover (e.g. stackedData changes). Without
-    // this, useChartDraw's hover effect re-fires on every drawHover identity change,
-    // and the resulting requestAnimationFrame churn can race with tooltip rendering.
+    // drawHover held via a ref so composedDrawHover stays referentially stable across
+    // parent drawHover identity changes (e.g. stackedData) — otherwise useChartDraw's
+    // hover effect refires on every render and the rAF churn races with tooltip rendering.
     const drawHoverRef = useRef(drawHover)
     drawHoverRef.current = drawHover
-    const composedDrawHover = useCallback(
-        (args: ChartDrawArgs) => {
-            if (showCrosshair && theme.crosshairColor && args.hoverIndex >= 0) {
-                const x = args.scales.x(args.labels[args.hoverIndex])
-                if (x != null && isFinite(x)) {
-                    drawCrosshair(args.ctx, args.dimensions, x, theme.crosshairColor)
-                }
-            }
-            drawHoverRef.current(args)
-        },
+    const composedDrawHover = useMemo(
+        () => composeDrawHoverWithCrosshair(() => drawHoverRef.current, theme.crosshairColor, showCrosshair),
         [showCrosshair, theme.crosshairColor]
     )
 
@@ -275,25 +179,7 @@ export function Chart<Meta = unknown>({
         [canvasRef]
     )
 
-    // Wrap resolveValue in a ref + stable callback so callers don't have to memoize it.
-    // An un-memoized arrow literal from a parent would otherwise invalidate the layout
-    // context on every render and defeat the layout/hover split.
-    //
-    // The ref is written during render rather than via an effect because overlays read
-    // it during their render via `useChartLayout().resolveValue`; deferring the write
-    // would expose them to last-commit's value. This is safe under StrictMode/concurrent
-    // rendering: an aborted render that wrote the ref will be re-driven with the same
-    // props, so the kept render observes an idempotent state.
-    const resolveValueRef = useRef<ResolveValueFn | undefined>(resolveValue)
-    resolveValueRef.current = resolveValue
-    const stableResolveValue = useCallback<ResolveValueFn>((s, i) => {
-        const fn = resolveValueRef.current
-        if (fn) {
-            return fn(s, i)
-        }
-        const v = s.data[i]
-        return typeof v === 'number' && Number.isFinite(v) ? v : 0
-    }, [])
+    const stableResolveValue = useStableResolveValue(resolveValue)
 
     const layoutValue = useMemo<ChartLayoutContextValue | null>(() => {
         if (!scales || !dimensions) {
